requirements.txt
asgiref==3.2.10
Django==3.1.1
django-froala-editor==3.2.2
django-json-field==0.5.7
django-jsonfield==1.4.0
djangorestframework==3.11.1
pytz==2020.5
sqlparse==0.3.1
waitress==2.0.0
whitenoise==5.2.0
requests==2.22.0
djangorestframework==3.11.1
psycopg2==2.8.6
sendgrid==6.4.8